Fix: renamed CI to confidence interval
